backend/requirements.txt
flask 
flask_sqlalchemy 
flask_mail
flask-mysqldb
mysql-connector-python